+++ app/assets/stylesheets/components/_navbar.scss
@@ -14,3 +14,9 @@
 .navbar-league-box .navbar-brand img {
   width: 40px;
 }
+
+.time-padding {
+  margin: auto;
+  padding-top: 6px;
+  padding-bottom: 6px;
+}
