+++ app/assets/stylesheets/config/_colors.scss
@@ -8,3 +8,6 @@ $orange: #E67E22;
 $green: #1EDD88;
 $gray: #0E0000;
 $light-gray: #F4F4F4;
+$tennis-green: #9cb879;
+$tennis-blue: #336699;
+$tennis-aqua: #03897c;
