+++ app/assets/stylesheets/pages/_boxes.scss
@@ -55,18 +55,20 @@
 }
 
 .box {
-  background-color: beige;
+  // background-color: beige;
+  background-color: $tennis-green;
   position: relative;
   justify-content: center;
   align-items: center;
-  border-radius: 15px;
+  border-radius: 0.25rem;
   padding: 5px;
   box-shadow: 0 0 15px rgba(0,0,0,0.2);
 
   &:hover {
-    transform: scale(1.02);
-    color: rgb(246, 137, 79);
-    border: solid 1px lightgrey;
+    transform: scale(1.01);
+    // color: rgb(246, 137, 79);
+    color: $tennis-blue;
+    // border: solid 1px lightgrey;
     box-shadow: 0 0 15px 5px rgba(0,0,0,0.1);
   }
 }
