+++ app/controllers/pages_controller.rb
@@ -1,5 +1,5 @@
 class PagesController < ApplicationController
-  skip_before_action :authenticate_user!, only: [:home, :overview, :structure, :staff]
+  skip_before_action :authenticate_user!, only: [:home, :rules, :structure, :staff]
 
   def staff
     if current_user == @admin
