+++ app/javascript/application.js
@@ -2,3 +2,5 @@
 import "@hotwired/turbo-rails"
 import "./controllers"
 import "bootstrap"
+import "@popperjs/core"
+
