+++ app/views/devise/passwords/edit.html.erb
@@ -21,7 +21,7 @@
     </div>
 
     <div class="form-actions flex-display">
-      <%= f.button :submit, "Change my password", class: "btn button-flat" %>
+      <%= f.button :submit, "Change my password", class: "btn buttons-shape btn-green" %>
     </div>
   <% end %>
