+++ app/views/devise/passwords/new.html.erb
@@ -12,7 +12,7 @@
     </div>
 
     <div class="form-actions flex-display">
-      <%= f.button :submit, "Send me reset password instructions", class: "btn button-flat" %>
+      <%= f.button :submit, "Send me reset password instructions", class: "btn buttons-shape btn-green" %>
     </div>
   <% end %>
