+++ app/views/devise/registrations/new.html.erb
@@ -35,7 +35,7 @@
     </div>
 
     <div class="form-actions flex-display">
-      <%= f.button :submit, "Sign up", class: "btn button-flat" %>
+      <%= f.button :submit, "Sign up", class: "btn buttons-shape btn-green" %>
     </div>
   <% end %>
