+++ app/views/devise/sessions/new.html.erb
@@ -14,7 +14,7 @@
     </div>
 
     <div class="form-actions flex-display">
-      <%= f.button :submit, "Log in", class: "btn button-flat"  %>
+      <%= f.button :submit, "Log in", class: "btn buttons-shape btn-green"  %>
     </div>
   <% end %>
