+++ app/views/matches/new.html.erb
@@ -1,10 +1,25 @@
 <div class = "container">
+  <br>
+  <h3>New match score</h3>
+
   <br><%= render "shared/club_round", round: @round %>
-  <h5><%= render "shared/fullname", user: @current_player %> against <%= render "shared/fullname", user: @opponent %></h5>
   <% players = [@current_player, @opponent] %>
   <br>
+  <div class="same-line">
+    <div>
+      <h5>
+        <%= render "shared/fullname", user: @current_player %><%= " vs " %>
+        <%= render "shared/fullname", user: @opponent %>
+      </h5>
+    </div>
+    <div>
+      <%= "Enter the new match score" %>
+    </div>
+    <div></div>
+  </div>
+
   <%= form_for @match do |f| %>
-    <div class = "form-wrap">
+    <div class = "frame-wrap">
       <div class="row">
         <div class="col-sm-2"><%= f.label :court_id, "Match played on court: " %></div>
         <div class="col-sm-1"><%= f.select :court_id, 1..@round.club.courts.count, default: 1 %></div>
@@ -41,12 +56,12 @@
     </div>
     <br>
     <div class="buttons-wrap">
-      <%= f.submit "Save score", class: "btn button-flat" %>
+      <%= f.submit "Save score", class: "btn buttons-shape btn-green" %>
       &nbsp;&nbsp;&nbsp;&nbsp;
       <% if current_user.role == "player" %>
-        <%= link_to "Back", manage_my_box_path(@box, page_from: manage_my_box_path(@box)), class: "btn button-flat-grey" %>
+        <%= link_to "Back", manage_my_box_path(@box, page_from: manage_my_box_path(@box)), class: "btn buttons-shape btn-blue" %>
       <% else %>
-        <%= link_to "Back", box_referee_path(@box, page_from: box_referee_path(@box)), class: "btn button-flat-grey" %>
+        <%= link_to "Back", box_referee_path(@box, page_from: box_referee_path(@box)), class: "btn buttons-shape btn-blue" %>
       <% end %>
     </div>
   <% end %>
