+++ app/views/matches/show.html.erb
@@ -5,32 +5,34 @@
   <h3><%= render "shared/fullname", user: current_user %><%= "(# #{player_ubs.rank})" %> vs
   <%= render "shared/fullname", user: @opponent %><%= "(# #{opponent_ubs.rank})" %></h3>
 
-  <div class="form-wrap">
+  <div class="frame-wrap">
     <br><h4><%= render "match", my_card: @current_user_match_score %>by <%= render "shared/fullname", user: current_user %></h4>
-    <br><%= "Played on #{@match.time.strftime("%d/%m/%Y")} (#{@match.time.strftime("%I:%M %p")}), court #{@match.court.name}." %>
+    <br><%= "Played on #{@match.time.strftime("%a %-d %h %y")} at #{@match.time.strftime("%H:%M")}, on court #{@match.court.name}." %>
 
     <br><br>
-    <%= "Score : " %><%= (render "shared/score", user_score: @current_user_match_score, opponent_score: @opponent_match_score).strip %>,
+    <strong><%= "Score : " %></strong>
+    <%= (render "shared/score", user_score: @current_user_match_score, opponent_score: @opponent_match_score).strip %>,
     <% ums =  UserMatchScore.find_by(match_id: @match.id, user_id: current_user.id) %>
     <% score_input_by = User.find(ums.input_user_id) %>
-    <% message = "submitted on #{ums.input_date.strftime("%d %b %Y")} #{ums.input_date.strftime("%I:%M %p")} by #{score_input_by.role} #{(render "shared/fullname", user: score_input_by).strip}." %>
-    <span style="color:grey"><%= message %></span>
+    <% message = "submitted on #{ums.input_date.strftime("%a %-d %h %y")} at #{ums.input_date.strftime("%H:%M")} by #{score_input_by.role} #{(render "shared/fullname", user: score_input_by).strip}." %>
+    <span style="color:#336699"><%= message %></span>
     <br><br>
-    <%= "Earnings : " %>
+    <strong><%= "Points earned : " %></strong>
     <%= (render "shared/fullname", user: current_user).strip %>
-    <%= ": #{pluralize @current_user_match_score.points, "point"} (total : #{pluralize player_ubs.points, "pt"}) - " %>
+    <%= ": #{pluralize @current_user_match_score.points, "point"} (#{pluralize player_ubs.points, "box pt"}) - " %>
     <%= (render "shared/fullname", user: @opponent).strip %>
-    <%= ": #{pluralize @opponent_match_score.points, "point"} (total : #{pluralize opponent_ubs.points, "pt"})." %>
+    <%= ": #{pluralize @opponent_match_score.points, "point"} (#{pluralize opponent_ubs.points, "box pt"})." %>
     <br><br>
     <% name = "#{render "shared/fullname", user: @referee}" %>
     <% message = "📞 #{@referee.phone_number}<br />✉️#{@referee.email}" %>
     <div data-controller="popover" data-bs-html="true" data-bs-toggle="popover" data-bs-trigger="hover" data-bs-placement="right" title="<%= name %>" data-bs-content="<%= message %>">
-      To challenge this match score, please contact your club referee <%= (render "shared/fullname", user: @referee).strip %>.
+      To challenge / edit this match score, please contact your club referee <%= (render "shared/fullname", user: @referee).strip %>.
     </div>
     <br>
   </div>
   <br>
   <div class="buttons-wrap">
-    <%= link_to "Back to my box", manage_my_box_path(@my_current_box), class: "btn button-flat-grey" %>
+    <%# link_to "Back to my box", manage_my_box_path(@my_current_box), class: "btn buttons-shape btn-blue" %>
+    <%= link_to "Back to my box", manage_my_box_path(@match.box), class: "btn buttons-shape btn-blue" %>
   </div>
 </div>
