+++ app/views/shared/_navbar.html.erb
@@ -12,31 +12,30 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav me-auto">
         <% if user_signed_in? %>
-          <li class="nav-item text-padding">
-            <%= "#{Time.now.strftime("%d %b %y")} #{Time.now.strftime("%H:%M")}" %>
+          <li class="nav-item active time-padding">
+            <%= "#{Time.now.strftime("%a %-d %b")} #{Time.now.strftime("%H:%M")}" %>&nbsp;&nbsp
           </li>
-          <li class="nav-item text-padding">
+          <li class="nav-item time-padding">
             <% if user_signed_in? && @round %>
               <%= " Round: #{@round.start_date.strftime("%b %y")} - #{@round.end_date.strftime("%b %y")}" %>
             <% end %>
           </li>
-          <li class="nav-item active">
+          <li class="nav-item">
             <%= link_to "Home", root_path, class: "nav-link" %>
           </li>
           <li class="nav-item">
             <% if current_user && current_user.role == "player" %>
               <%# passing 0 to manage_my_box_path, forces user to chose a round%>
-              <%= link_to "My box results", manage_my_box_path(0), class: "nav-link" %>
+              <%= link_to "My box", manage_my_box_path(0), class: "nav-link" %>
             <% end %>
           </li>
-          <li class="nav-item">
-            <%= link_to "All boxes", boxes_path, class: "nav-link" %>
-          </li>
-          <li class="nav-item">
-            <%= link_to "League table", user_box_scores_path, class: "nav-link" %>
-          </li>
-          <li class="nav-item">
-            <%= link_to "Log out (#{current_user.first_name} #{current_user.last_name} - #{current_user.role})", destroy_user_session_path, data: {turbo_method: :delete}, class: "nav-link" %>
+          <li class="nav-item dropdown">
+            <a class="nav-link" href="#" data-bs-toggle="dropdown"><%= "#{current_user.nickname} - #{current_user.role}" %></a>
+            <div class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
+              <%= link_to "All boxes", boxes_path, class: "dropdown-item" %>
+              <%= link_to "League table", user_box_scores_path, class: "dropdown-item" %>
+              <%= link_to "Log out", destroy_user_session_path, data: {turbo_method: :delete}, class: "dropdown-item" %>
+            </div>
           </li>
         <% else %>
           <li class="nav-item">
