+++ app/views/shared/_round.html.erb
@@ -1,2 +1,2 @@
 <%# partial: display round start date and end date %>
-<%= "Round : #{round.start_date.strftime("%d %b %Y")} - #{round.end_date.strftime("%d %b %Y")}" %>
+<%= "Round : #{round.start_date.strftime("%d %b")} to #{round.end_date.strftime("%d %b %y")}" %>
