+++ app/views/shared/_set_club_and_round.html.erb
@@ -6,8 +6,8 @@
       <%= form.label :club_name, "Select a club: " %>
       <%= form.select :club_name, @club_names, :selected => params[:club_name] %>
       <% if !params[:club_name] || params[:round_start] %>
-        &nbsp;&nbsp;&nbsp;&nbsp;<%= form.submit "Confirm", class: "btn button-flat" %>
-        &nbsp;&nbsp;&nbsp;&nbsp;<%= link_to "Back", "javascript:history.back()", class: "btn button-flat-grey" %>
+        &nbsp;&nbsp;&nbsp;&nbsp;<%= form.submit "Confirm", class: "btn buttons-shape btn-green" %>
+        &nbsp;&nbsp;&nbsp;&nbsp;<%= link_to "Back", "javascript:history.back()", class: "btn buttons-shape btn-blue" %>
       <% end %>
     <% end %>
   <% end %>
@@ -24,7 +24,7 @@
     <%= form.label :round_start, "Select a round start date: " %>
     <%= form.select :round_start, @start_dates, :selected => params[:round_start] %>&nbsp;&nbsp;&nbsp;&nbsp;
     <%= hidden_field_tag(:club_name, @club.name) %>
-    <%= form.submit "Confirm", class: "btn button-flat" %>
-    &nbsp;&nbsp;&nbsp;&nbsp;<%= link_to "Back", "javascript:history.back()", class: "btn button-flat-grey" %>
+    <%= form.submit "Confirm", class: "btn buttons-shape btn-green" %>
+    &nbsp;&nbsp;&nbsp;&nbsp;<%= link_to "Back", "javascript:history.back()", class: "btn buttons-shape btn-blue" %>
   <% end %>
 <% end %>
