+++ app/views/user_box_scores/new.html.erb
@@ -2,7 +2,7 @@
 <%# the csv file must contain fields id, email, first_name, last_name, (nickname), phone_number, role (players + 1 referee) %>
 <%# players are allocated in boxes by id in descending order  %>
 <div class="container">
-  <h3>Create a new club</h3><br>
+  <h3>Create a new club and a round</h3><br>
   <% unless current_user.role == "admin" %>
     <br><br><%= "You are not an admin, this page is not authorised." %>
   <% else %>
@@ -12,7 +12,7 @@
     <p><i>id, email, first_name, last_name, [nickname], phone_number, role</i>
     . Players will be allocated in boxes based on the id order.</p>
     <%= form_with(url: {action: "create"}, multipart: true) do %>
-      <div class="form-wrap">
+      <div class="frame-wrap">
         <div class="row"><br>
           <div class="col-sm-2"><%= label_tag :new_club_name %></div>
           <div class="col-sm-1"><%= text_field_tag :new_club_name, "My tennis club ltc", size: 30 %></div>
@@ -43,8 +43,8 @@
         </div><br>
       </div><br>
       <div class="buttons-wrap">
-        <%= submit_tag 'Submit', class: "btn button-flat" %>
-        <%= link_to "Back", "javascript:history.back()", class: "btn button-flat-grey" %>
+        <%= submit_tag 'Submit', class: "btn buttons-shape btn-green" %>
+        <%= link_to "Back", "javascript:history.back()", class: "btn buttons-shape btn-blue" %>
       </div>
     <% end %>
   <% end %>
