+++ config/initializers/default_meta.rb
@@ -1,0 +1,1 @@
+DEFAULT_META = YAML.load_file(Rails.root.join("config/meta.yml"))
