+++ config/meta.yml
@@ -1,0 +1,3 @@
+meta_name: "LeagueBox"
+meta_title: "LeagueBox · The path to social tennis playing in Belgium"
+meta_description: "Join your box league to play social tennis in your club"
