+++ package.json
@@ -5,7 +5,7 @@
     "@hotwired/stimulus": "^3.2.1",
     "@hotwired/turbo-rails": "^7.3.0",
     "@popperjs/core": "^2.11.7",
-    "bootstrap": "^5.2.3",
+    "bootstrap": "^5.3.2",
     "webpack": "^5.77.0",
     "webpack-cli": "^5.0.1"
   },
