+++ public/500.html
@@ -4,63 +4,74 @@
   <title>We're sorry, but something went wrong (500)</title>
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <style>
-  .rails-default-error-page {
-    background-color: #EFEFEF;
-    color: #2E2F30;
+  /* custom CSS */
+  .container {
+    margin: 20rem auto;
+    width: 50%;
+  }
+
+  .message {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    align-content: space-around;
+
+  }
+  .frame-wrap {
     text-align: center;
-    font-family: arial, sans-serif;
-    margin: 0;
+    background-color: #9cb879;
+    margin: 0 auto;
+    padding: 5px 24px;
+    border-radius: 0.5rem;
+    box-shadow: 0 0 8px rgba(0,0,0,0.2);
   }
 
-  .rails-default-error-page div.dialog {
-    width: 95%;
-    max-width: 33em;
-    margin: 4em auto 0;
+  .buttons-wrap {
+  display: flex;
+  padding: 5px 24px;
+  flex-wrap: wrap;
   }
 
-  .rails-default-error-page div.dialog > div {
-    border: 1px solid #CCC;
-    border-right-color: #999;
-    border-left-color: #999;
-    border-bottom-color: #BBB;
-    border-top: #B00100 solid 4px;
-    border-top-left-radius: 9px;
-    border-top-right-radius: 9px;
-    background-color: white;
-    padding: 7px 12% 0;
-    box-shadow: 0 3px 8px rgba(50, 50, 50, 0.17);
+  .button-blue {
+  margin: 0 auto 5px;
+  color: white;
+  padding: 6px 24px;
+  border-radius: 0.25rem;
+  text-decoration: none;
+  background: #336699;
+  transition: background 0.3s ease;
   }
 
-  .rails-default-error-page h1 {
-    font-size: 100%;
-    color: #730E15;
-    line-height: 1.5em;
+  .button-blue:hover {
+    background: #7f7f7f;
+    color: white;
   }
-
-  .rails-default-error-page div.dialog > p {
-    margin: 0 0 1em;
-    padding: 1em;
-    background-color: #F7F7F7;
-    border: 1px solid #CCC;
-    border-right-color: #999;
-    border-left-color: #999;
-    border-bottom-color: #999;
-    border-bottom-left-radius: 4px;
-    border-bottom-right-radius: 4px;
-    border-top-color: #DADADA;
-    color: #666;
-    box-shadow: 0 3px 8px rgba(50, 50, 50, 0.17);
-  }
-  </style>
+</style>
 </head>
 
-<body class="rails-default-error-page">
+<body>
   <!-- This file lives in public/500.html -->
-  <div class="dialog">
-    <div>
-      <h1>We're sorry, but something went wrong.</h1>
+  <div class="container">
+    <div class="message">
+      <div>
+        <img src="tennis_racket.jpg" alt="tennis_racket image"  height=50>
+      </div>
+      <div class="frame-wrap">
+        <h1>League box</h1>
+        <h3>We're sorry, but something went wrong.</h3>
+        <p>If you are the application owner check the logs for more information.</p>
+          </div>
+      <div>
+        <img  src="tennis_racket.jpg" alt="tennis_racket image"  height=50>
+      </div>
     </div>
-    <p>If you are the application owner check the logs for more information.</p>
+    <br>
+    <div class = "buttons-wrap">
+      <a  class="button-blue" href="/">Home</a>
+    </div>
   </div>
+
+
 </body>
+
 </html>
